calculos/requirements.txt
Flask
pymongo
psycopg2-binary
